QA-8315: re-moved EServiceAttributesKey.java and EServiceAttributesGroupKey.java
